requirements.txt
slackclient>=2.0.0
certifi
mysql-connector-python
pandas
psycopg2-binary
slackeventsapi
scipy
python-crontab
requests
